Move IsSerialInBlockchain to zerocoin_verify.h/cpp
And, as it's only used inside this function, merge it with the IsTransactionInChain function.
