+++ src/init.cpp
@@ -55,7 +55,6 @@
 #include "util/threadnames.h"
 #include "validation.h"
 #include "validationinterface.h"
-#include "zpivchain.h"
 #include "warnings.h"
 
 #ifdef ENABLE_WALLET
